+++ src/utils/common/params.h
@@ -129,9 +129,9 @@ void params_flag_verbose(params_t *params);
 
 int params_parse_port(const char *value, char **port);
 
-int params_parse_class(const char *value, int32_t *rclass);
+int params_parse_class(const char *value, uint16_t *rclass);
 
-int params_parse_type(const char *value, int32_t *rtype, uint32_t *xfr_serial);
+int params_parse_type(const char *value, uint16_t *rtype, uint32_t *xfr_serial);
 
 int params_parse_server(const char *value, list *servers, const char *def_port);
